requirements/common.txt
python-telegram-bot==21.10
pydub==0.25.1
faster-whisper==1.1.1
pydantic-settings==2.7.1
